+++ RELEASENOTES.md
@@ -2,6 +2,8 @@
 
 ### Next version
 - SD card usage in %
+- new confirm message dialog
+- Memory usage menu item
 
 ### v2.6.0 2021-12-27
 - FIX added PSRAM option for M5 Core2 and TTGO T4 (BLE4+SD+Wifi is now working)